Refactor & upgrade react-scripts package
