--- src/utils/image-id.js
@@ -1,11 +1,0 @@
-export const imageSrcToPictureId = src => {
-  if (src.includes("data:image")) return "default"
-  let id
-  id = src
-    .split("\\")
-    .pop()
-    .split("/")
-    .pop() // get rid of domain and pathname
-  id = id.replace(/\.[^/.]+$/, "") // get rid of extension
-  return id
-}
